Cleaned up to code and separated it into different files
