+++ AI-dvisor.xcodeproj/xcuserdata/cmj3292.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -36,7 +36,21 @@
             landmarkType = "9">
             <Locations>
                <Location
-                  uuid = "EDAEA565-8997-45A8-B57A-BF5C9AECF2E9 - 6d12f269a8de55a"
+                  uuid = "EDAEA565-8997-45A8-B57A-BF5C9AECF2E9 - 448e1c66fa34c047"
+                  shouldBeEnabled = "Yes"
+                  ignoreCount = "0"
+                  continueAfterRunningActions = "No"
+                  symbolName = "saveTextAsPDF(pdfContent: Swift.String, fileName: Swift.String) -&gt; ()"
+                  moduleName = "AI-dvisor.debug.dylib"
+                  usesParentBreakpointCondition = "Yes"
+                  urlString = "file:///Users/cmj3292/Desktop/AI-dvisor-iOS-Project/AI-dvisor/Chatbot/ChatbotViewController.swift"
+                  startingColumnNumber = "9223372036854775807"
+                  endingColumnNumber = "9223372036854775807"
+                  startingLineNumber = "250"
+                  endingLineNumber = "250">
+               </Location>
+               <Location
+                  uuid = "EDAEA565-8997-45A8-B57A-BF5C9AECF2E9 - 6d12f269a8de51c"
                   shouldBeEnabled = "Yes"
                   ignoreCount = "0"
                   continueAfterRunningActions = "No"
@@ -46,22 +60,8 @@
                   urlString = "file:///Users/cmj3292/Desktop/AI-dvisor-iOS-Project/AI-dvisor/Chatbot/ChatbotViewController.swift"
                   startingColumnNumber = "9223372036854775807"
                   endingColumnNumber = "9223372036854775807"
-                  startingLineNumber = "249"
-                  endingLineNumber = "249">
-               </Location>
-               <Location
-                  uuid = "EDAEA565-8997-45A8-B57A-BF5C9AECF2E9 - 4ee5144774e9629d"
-                  shouldBeEnabled = "Yes"
-                  ignoreCount = "0"
-                  continueAfterRunningActions = "No"
-                  symbolName = "closure #1 @Swift.MainActor () -&gt; () in closure #2 (Swift.String) -&gt; () in saveTextAsPDF(pdfContent: Swift.String, fileName: Swift.String) -&gt; ()"
-                  moduleName = "AI-dvisor.debug.dylib"
-                  usesParentBreakpointCondition = "Yes"
-                  urlString = "file:///Users/cmj3292/Desktop/AI-dvisor-iOS-Project/AI-dvisor/Chatbot/ChatbotViewController.swift"
-                  startingColumnNumber = "9223372036854775807"
-                  endingColumnNumber = "9223372036854775807"
-                  startingLineNumber = "250"
-                  endingLineNumber = "250">
+                  startingLineNumber = "251"
+                  endingLineNumber = "251">
                </Location>
             </Locations>
          </BreakpointContent>
