+++ AI-dvisor/HomeScreen/HomeScreenViewController.swift
@@ -11,7 +11,6 @@ import FirebaseAuth
 
 // This extension allows the HomeScreenViewController to handle new journal creation updates.
 extension HomeScreenViewController: NewJournalDelegate, EditJournalDelegate, AddEventDelegate, AddTimerDelegate {
-
     
     func didCreateJournal(_ journal: Journal) {
         journals.append(journal)
@@ -72,46 +71,46 @@ class HomeScreenViewController: UIViewController, UICollectionViewDataSource, UI
         let longPressGesture = UILongPressGestureRecognizer(target: self, action: #selector(handleLongPress))
         journalCollectionView.addGestureRecognizer(longPressGesture)
         
-        // So teh SegCtrl text color can be different
+        // So the SegCtrl text color can be different
         let selectedTextAttributes: [NSAttributedString.Key: Any] = [
             .foregroundColor: UIColor.white  // Color for selected text
         ]
-
+        
         let unselectedTextAttributes: [NSAttributedString.Key: Any] = [
             .foregroundColor: UIColor.black  // Color for unselected text
         ]
-
+        
         sortJournalSegCtrl.setTitleTextAttributes(selectedTextAttributes, for: .selected)
         sortJournalSegCtrl.setTitleTextAttributes(unselectedTextAttributes, for: .normal)
     }
-
+    
     private func fetchCurrentUser() {
-          guard let email = Auth.auth().currentUser?.email else {
+        guard let email = Auth.auth().currentUser?.email else {
             fatalError("No Firebase user logged in")
-          }
-
-          let req: NSFetchRequest<User> = User.fetchRequest()
-          req.predicate = NSPredicate(format: "email == %@", email)
-
-          do {
+        }
+        
+        let req: NSFetchRequest<User> = User.fetchRequest()
+        req.predicate = NSPredicate(format: "email == %@", email)
+        
+        do {
             if let user = try context.fetch(req).first {
-              // found an existing record
-              let isDarkMode = user.value(forKey: "darkMode") as? Bool ?? false
-              onDarkLightMode(darkMode: isDarkMode)
-              currentUser = user
+                // found an existing record
+                let isDarkMode = user.value(forKey: "darkMode") as? Bool ?? false
+                onDarkLightMode(darkMode: isDarkMode)
+                currentUser = user
             } else {
-              // **not** found — create one now
-              let newUser = User(context: context)
-              newUser.email       = email
-              newUser.username    = email.components(separatedBy: "@").first
-              // you can set a default profile picture here if you want
-              try context.save()
-              currentUser = newUser
-              print("Created Core Data User for \(email)")
+                // **not** found — create one now
+                let newUser = User(context: context)
+                newUser.email       = email
+                newUser.username    = email.components(separatedBy: "@").first
+                // you can set a default profile picture here if you want
+                try context.save()
+                currentUser = newUser
+                print("Created Core Data User for \(email)")
             }
-          } catch {
+        } catch {
             fatalError("Failed fetching/creating User: \(error)")
-          }
+        }
     }
     
     // For Dark/Light Mode
@@ -125,10 +124,10 @@ class HomeScreenViewController: UIViewController, UICollectionViewDataSource, UI
             addFriend.tintColor = UIColor(red: 211/255.0, green: 219/255.0, blue:  178/255.0,alpha: 1.0)
             addDueDate.tintColor = UIColor(red: 211/255.0, green: 219/255.0, blue:  178/255.0,alpha: 1.0)
         } else {
-           // Light Mode:
+            // Light Mode:
             view.backgroundColor = UIColor(red: 211/255.0, green: 219/255.0, blue:  178/255.0,alpha: 1.0)
             journalCollectionView.backgroundColor = UIColor(red: 211/255.0, green: 219/255.0, blue:  178/255.0,alpha: 1.0)
-
+            
         }
     }
     
@@ -136,7 +135,7 @@ class HomeScreenViewController: UIViewController, UICollectionViewDataSource, UI
         super.viewWillAppear(animated)
         loadJournals()
         journalCollectionView.reloadData()
-      }
+    }
     
     @IBAction func sortJournals(_ sender: Any) {
         loadJournals()
@@ -145,70 +144,70 @@ class HomeScreenViewController: UIViewController, UICollectionViewDataSource, UI
     private func loadJournals() {
         
         // Ensure we have the logged‑in User
-            guard let user = currentUser else {
-                cdJournals = []
-                journals   = []
-                journalCollectionView.reloadData()
-                return
-            }
-
-            // Build & run the fetch request
-            let req: NSFetchRequest<UserJournal> = UserJournal.fetchRequest()
-            req.predicate = NSPredicate(format: "users == %@", user)
-        
-        
-            do {
-                // Fetch the managed objects
-                cdJournals = try context.fetch(req)
-                
-                // Sort based on which segment of the segctrl is selected
-                switch sortJournalSegCtrl.selectedSegmentIndex {
-                case 0:
-                    // alphabetical by title
-                    cdJournals.sort {
-                                   ($0.title ?? "").localizedCaseInsensitiveCompare($1.title ?? "") == .orderedAscending
-                               }
-                case 1:
-                    cdJournals.sort {
-                                    ($0.importance?.count ?? 0) > ($1.importance?.count ?? 0)
-                                }
-                default:
-                    break
-                }
-
-                // Map each managed object into your Swift struct
-                journals = cdJournals.map { entity in
-                    let title      = entity.title      ?? "Untitled"
-                    let importance = entity.importance ?? "!"
-                    
-                    // Unarchive UIColor from Data, fallback to .blue
-                    let bgColor: UIColor
-                    if let data = entity.bgColor,
-                       let color = try? NSKeyedUnarchiver
-                                        .unarchivedObject(ofClass: UIColor.self, from: data) {
-                        bgColor = color
-                    } else {
-                        bgColor = .blue
-                    }
-
-                    return Journal(
-                        title:      title,
-                        importance: importance,
-                        bgColor:    bgColor
-                    )
-                }
-
-            } catch {
-                print("Failed to fetch journals:", error)
-                cdJournals = []
-                journals   = []
-            }
-
-            // Refresh the UI
+        guard let user = currentUser else {
+            cdJournals = []
+            journals   = []
             journalCollectionView.reloadData()
+            return
+        }
+        
+        // Build & run the fetch request
+        let req: NSFetchRequest<UserJournal> = UserJournal.fetchRequest()
+        req.predicate = NSPredicate(format: "users == %@", user)
+        
+        
+        do {
+            // Fetch the managed objects
+            cdJournals = try context.fetch(req)
+            
+            // Sort based on which segment of the segctrl is selected
+            switch sortJournalSegCtrl.selectedSegmentIndex {
+            case 0:
+                // alphabetical by title
+                cdJournals.sort {
+                    ($0.title ?? "").localizedCaseInsensitiveCompare($1.title ?? "") == .orderedAscending
+                }
+            case 1:
+                cdJournals.sort {
+                    ($0.importance?.count ?? 0) > ($1.importance?.count ?? 0)
+                }
+            default:
+                break
+            }
+            
+            // Map each managed object into your Swift struct
+            journals = cdJournals.map { entity in
+                let title      = entity.title      ?? "Untitled"
+                let importance = entity.importance ?? "!"
+                
+                // Unarchive UIColor from Data, fallback to .blue
+                let bgColor: UIColor
+                if let data = entity.bgColor,
+                   let color = try? NSKeyedUnarchiver
+                    .unarchivedObject(ofClass: UIColor.self, from: data) {
+                    bgColor = color
+                } else {
+                    bgColor = .blue
+                }
+                
+                return Journal(
+                    title:      title,
+                    importance: importance,
+                    bgColor:    bgColor
+                )
+            }
+            
+        } catch {
+            print("Failed to fetch journals:", error)
+            cdJournals = []
+            journals   = []
+        }
+        
+        // Refresh the UI
+        journalCollectionView.reloadData()
     }
     
-    // This function is not implemented yet for Alpha but will be implemented for Final
+    // This function acts just the same as a pressedButton function but for images.
     @objc func addFriendImageTapped(_ sender: UITapGestureRecognizer) {
         let storyboard = UIStoryboard(name: "HomeScreenStoryboard", bundle: nil)
         
@@ -221,6 +220,7 @@ class HomeScreenViewController: UIViewController, UICollectionViewDataSource, UI
         }
     }
     
+    // This function acts just the same as a pressedButton function but for images.
     @objc func addTimerImageTapped(_ sender: UITapGestureRecognizer) {
         let storyboard = UIStoryboard(name: "HomeScreenStoryboard", bundle: nil)
         
@@ -237,6 +237,7 @@ class HomeScreenViewController: UIViewController, UICollectionViewDataSource, UI
         }
     }
     
+    // This function acts just the same as a pressedButton function but for images.
     @objc func addDueDateTapped(_ sender: UITapGestureRecognizer) {
         let storyboard = UIStoryboard(name: "HomeScreenStoryboard", bundle: nil)
         
@@ -276,7 +277,7 @@ class HomeScreenViewController: UIViewController, UICollectionViewDataSource, UI
             let cell = journalCollectionView.dequeueReusableCell(withReuseIdentifier: "AddJournalCell", for: indexPath)
             cell.layer.cornerRadius = 15
             cell.layer.masksToBounds = true
-
+            
             return cell
         } else {
             let cell = journalCollectionView.dequeueReusableCell(withReuseIdentifier: "JournalCell", for: indexPath) as! JournalCollectionViewCell
@@ -285,12 +286,12 @@ class HomeScreenViewController: UIViewController, UICollectionViewDataSource, UI
             cell.journalView.backgroundColor = journals[indexPath.row - 1].bgColor
             
             cell.contentView.layer.cornerRadius = 15
-                    cell.contentView.layer.masksToBounds = true
-
-                    // Apply shadow
-                    cell.layer.shadowColor = UIColor.black.cgColor
-                    cell.layer.shadowOpacity = 0.4
-                    cell.layer.shadowOffset = CGSize(width: 2, height: 2)
+            cell.contentView.layer.masksToBounds = true
+            
+            // Apply shadow
+            cell.layer.shadowColor = UIColor.black.cgColor
+            cell.layer.shadowOpacity = 0.4
+            cell.layer.shadowOffset = CGSize(width: 2, height: 2)
             cell.layer.shadowRadius = 4
             cell.layer.masksToBounds = false
             return cell
@@ -371,8 +372,8 @@ class HomeScreenViewController: UIViewController, UICollectionViewDataSource, UI
             
             let selectedJournal = cdJournals[indexPath.item - 1]
             editJournalVC.userJournal = selectedJournal
-                    editJournalVC.journalIndex = indexPath.item - 1
-                    editJournalVC.currentUser   = currentUser
+            editJournalVC.journalIndex = indexPath.item - 1
+            editJournalVC.currentUser   = currentUser
             
             if let sheet = editJournalVC.sheetPresentationController {
                 sheet.detents = [.medium()] // Makes it take up half the screen
