+++ AI-dvisor/HomeScreen/Journal Stuff/JournalCollectionViewCell.swift
