+++ AI-dvisor/HomeScreen/Journal Stuff/UserJournal+CoreDataClass.swift
