+++ AI-dvisor/HomeScreen/Journal Stuff/UserJournal+CoreDataProperties.swift
