+++ AI-dvisor/HomeScreen/Timer Stuff/AddTimerViewController.swift
@@ -11,7 +11,7 @@ protocol AddTimerDelegate: AnyObject {
     func didAddTimer()
 }
 
-var pickerData = ["30 minutes","45 minutes","1 hour","1 hour 15 minutes","1 hour 30 minutes","1 hour 45 minutes","2 hours", "10 Seconds"]
+var pickerData = ["30 minutes","45 minutes","1 hour","1 hour 15 minutes","1 hour 30 minutes","1 hour 45 minutes","2 hours","10 Seconds"]
 
 class AddTimerViewController: UIViewController, UIPickerViewDelegate, UIPickerViewDataSource {
 
@@ -29,7 +29,6 @@ class AddTimerViewController: UIViewController, UIPickerViewDelegate, UIPickerVi
     override func viewDidLoad() {
         super.viewDidLoad()
 
-        // Do any additional setup after loading the view.
         timePicker.delegate = self
         timePicker.dataSource = self
         
@@ -58,6 +57,7 @@ class AddTimerViewController: UIViewController, UIPickerViewDelegate, UIPickerVi
         let selected = pickerData[row]
                 switch selected {
                 case "30 minutes":
+                    // Doing this math so it convert seconds to minutes becuase "Time Interval" is only in seconds
                     duration =  30*60
                     selectedRow = 0
                 case "45 minutes":
@@ -78,6 +78,7 @@ class AddTimerViewController: UIViewController, UIPickerViewDelegate, UIPickerVi
                 case "2 hours":
                     duration = 120*60
                     selectedRow = 6
+                    // This case was for the demo but I'm leaving it in for grading purposes
                 case "10 Seconds":
                     duration = 10
                     selectedRow = 7
@@ -87,26 +88,13 @@ class AddTimerViewController: UIViewController, UIPickerViewDelegate, UIPickerVi
     }
     
     @IBAction func setButtonTapped(_ sender: Any) {
-//        timer?.invalidate()
-//                
-//                timer = Timer.scheduledTimer(timeInterval: duration,
-//                                             target: self,
-//                                             selector: #selector(timerFinished),
-//                                             userInfo: nil,
-//                                             repeats: false)
+        // This line is so that the timer runs no matter what view you are on
+        // So that the alert will show up on any screen not just the that set it.
         TimerClass.shared.start(duration: duration)
-            delegate?.didAddTimer()
+        delegate?.didAddTimer()
         self.dismiss(animated: true)
-        
     }
     
-//    @objc private func timerFinished() {
-//            let alert = UIAlertController(title: "Time to take a Break!",
-//                                          message: "Your \(selectedRow) timer has finished.",
-//                                          preferredStyle: .alert)
-//            alert.addAction(.init(title: "OK", style: .default))
-//            present(alert, animated: true)
-//        }
     @IBAction func cancleButtonTapped(_ sender: Any) {
         self.dismiss(animated: true)
     }
